feat(region-map): add region_map_animals with encounter and wandering logic; wire into runtime
